split api in two file (create network.ts)
