+++ packages/auto-consensus/src/index.ts
@@ -1,3 +1,4 @@
 export * from './api'
 export * from './balances'
 export * from './info'
+export * from './network'
